Moved order row functions out of order item component.
